+++ CHANGELOG.md
@@ -8,6 +8,7 @@ nd this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.
 
 - Add cache support for DataLoader. [#455](https://github.com/async-graphql/async-graphql/issues/455)
 - Prevent Warp WS Close, Ping, and Pong messages from being parsed as GraphQL [#459](https://github.com/async-graphql/async-graphql/pull/459)
+- Implements `ScalarType` for `serde_json::Value`.
 
 ## [2.7.2]